+++ config/src/main/java/org/hyperledger/besu/config/StubGenesisConfigOptions.java
@@ -480,6 +480,11 @@ public class StubGenesisConfigOptions implements GenesisConfigOptions, Cloneable
     return Optional.empty();
   }
 
+  @Override
+  public Optional<BlobScheduleOptions> getBlobScheduleOptions() {
+    return Optional.empty();
+  }
+
   /**
    * Homestead block stub genesis config options.
    *